fix template class of CPT in java modular.
